backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
pycryptodome
supabase
slowapi
